requirements.txt
python-dotenv==1.0.1
requests==2.31.0
playwright==1.42.0
pandas==2.2.1
pydantic==2.6.3
schedule==1.2.1
loguru==0.7.2
instagram-private-api==1.6.0.0
pillow==10.2.0
aiohttp==3.9.3 